fee quoter cap issuance and small lints

+++ relayer/codec/encoder.go
@@ -140,9 +140,8 @@ func encodeUint(typeName string, value any) (any, error) {
 			if err != nil {
 				return nil, fmt.Errorf("cannot convert json.Number %s to %s: %w", v, typeName, err)
 			}
-			//nolint:gosec
 			// we assume safe conversion without negative numbers
-			baseValue = uint64(i)
+			baseValue = i
 		}
 	case string:
 		// Handle big numbers that might come as strings
